refactoring of ecnf utility code
